Implement Gradebook create() and save()
This introduces create() and save() methods inside of the Gradebook
class. The former is called once to create a new instance of the
Gradebook class while also generating metadata and ensuring the intended
filepath exists. The latter includes a helper write_json() method and
then writes the serialized core data (students, assignments,
submissions) to file.

This commit also includes an empty requirements.txt file and stubs for
the four other data classes.