Gradebook.add_student() and test coverage
This commit updates the Gradebook class to include add_student() method,
which takes a Student object and adds it the class record. A test suite
for Gradebook and Student classes is also frameworked, using pytest and
implementing coverage for methods as they are written. Not quite TDD,
but definitely moving slow enough to make sure that new features are
accompanied with corresponding tests.

Also adds __str__() and __repr__() methods to the Student class.