Core data models completed
This commit includes basic implemenations for the Gradebook, Student,
Assignment, Category, and Submission classes, as well as test coverage
for adding each to a gradebook, to_dict() and from_dict() methods, as
well as a few class specific things (late submission, category
weighting).

In Gradebook, the submissions attribute was changed from a list to a
dictionary, keyed by submission id, in order to better match the
students and assignments attributes.

Lastly, a utils file was started. It presently contains only a UUID
generator to be shared by the various ID identified classes.