Scaffolding the CLI menu loop
This commit includes ground work for the CLI menu loop. It adds a
run_cli() method in cli/main.py that acts as the application launch.
This commit also includes classes path_utils.py and menu_helpers.py
which, as the names suggest, provide functions to help reduce the
boilerplate in dealing with files and CLI menu prompts as this branch
goes along.