Creating and loading Gradebooks
This commit tackles the first rung of the CLI menu that the application
hooks into -- create_gradebook() and load_gradebook(). Create gradebook
takes a course name, course term, and optionally, a user defined
directory to store the Gradebook data (otherwise using a default
directory path). The method does detect whether the directory is empty
or not, and if not, warns the user that there is a possibiltiy of
overwriting existing data before confirming and proceeding.

The load_gradebook() method solicits the directory path from the user of
existing Gradebook data, and if it exists, passes the directory path off
to Gradebook.load(). This class method handles the deserializing of the
JSON data files and storing them into class attributes. It is
essentially the from_dict() method for Gradebooks.

Both the create and save gradebook options return a Gradebook object to
the run_cli() method, which will then hand the gradebook off to the
other menus.