Add Course Menu
This adds the first major CLI submenu, just beneath the create/load
gradebook layer. The course menu is the first layer for manipulating a
grade book object. This version currently only stores stubs for each
dispatch, except for the option to save the gradebook.

Gradebook class has been updated to include a private attribute
_dir_path which is written each time a gradebook object is created or
loaded from memory. This is not stored in the gradebook.metadata (since
both the metadata.json and/or the most recent save directory can be
manipulated outside of the program), but rather written each time to the
object instance upon creation. This is to help ensure that any calls to
save write to the most recent location that data was read from, unless
explicitly stated otherwise (i.e. "save as").

The menu helper display_menu was also updated to include an optional
zero_option to allow for finer control over the "0. Return" option
tagged to each menu.