Adding and searching for students
This implements first attempts at two more important functions of the
students_menu CLI: adding students and searching for students. Adding
students is fairly rudimentary at this point. Solicit input from the
user for first name, last name, and email address. All three must be
entered to proceed, but otherwise there is no input validation yet. UUID
is generated and a new Student object is instantiated and added to the
gradebook.students. Confirmation is printed.

Searching for students is only slightly more complicated, but largely
because it involved the creation of a few helper functions that will be
used in other corners of the program. In menu_helpers.py,
display_results takes an iterable, and optionally, a boolean flag
show_index and a formatter function. It displays the results line by
line, with or without numbering and according to a given formatter
function (print(str(x)) by default). Back in students_menu, the
search_students() prompts the user for a query, and then returns a list
of any students whose full name or email contains the query term.
view_student begins by calling search_students() and passes the search
results to prompt_student_selection(). This helper method takes a list
of students and if the list is empty, prints a "No hits" message; if the
list has one item, returns only that object; and if the list is 2 or
more items, announces how many items were found, displays the results
(using display_results), and prompts the user to choose the student they
wish to view by entering their index number (displayed on screen).

Two small other notes. A simple helper formatter function is defined
separately as DRY hygiene. Also, the Student class was refactored a bit
to have separate first name and last name attributes. We were headed
that way ultimately anyway, and it was clear that it would be more of a
headache to change it later down the road.