Changes to display_menu
I decided to refactor display_menu in two significant ways. First of
all, the options parameter is now a list of tuples rather than a
dictionary, since I was silently relying or ordered dictionaries and it
there was a risky reliance on the order of .keys() and .values() being
maintained across the call. The list is safer, and I wasn't so far into
menu writing that it was laborious to refactor. Secondly, the
display_menu() doesn't actual execute anything, it only returns an
action (or MenuSignal), and it is up to the caller to take action. This
proved necessary for when the functions stored as options don't yet have
access to some parameter. By returning just the function to be called as
the result of the display_menu, the caller to display_menu can now
supply the argument that was otherwise inaccessible. All in all, not too
painful to refactor and well worth the trouble for making this helper
function actually helpful.