Formatting and whitespace changes to student_menu
Combed through all menu interactions in the Manage Students menu to make
the line spacing and menu formatting as consistent as possible. Updated
the user input requests to use a helper function prompt_user_input()
that visually distinguishes when the program is waiting on the user.
Banner formatting is stylized and made consistent through another helper
function. And the edit student details is made more uniform, as well as
including a "do you want to keep editing this record" allowing the user
to stay in the record or kick all the way back up to the Manage Students
menu.