Minor changes to students_menu
Adds a slightly modified confirmation upon adding students to the
roster, and also improves the student creation function to include a
try/except block on object initialization. This is to establish the
pattern that can gracefully handle data validation in the class property
setter methods that have yet to be written.