Implement categories_menu and update Category
The category class is updated to include several property and setter
methods, particularly with validity checks for category weights. Methods
for reporting and toggling a category's archived/active status are
included.

The Manage Categories menu is also introduced in this commit. It closely
follows the Manage Students menu patterns, with means for adding,
editing, removing, viewing, and searching for categories. Viewing
includes means for viewing active, archived, or all categories. There
are methods for deleting, archiving, and reactivating categories. And
lastly, includes a dispatch command for calling weights_menu.run, the
sub-menu responsible for activating and manipulating category weights.