Optional, DRY helpers, and archived statuses
This commit is a big grab of housekeeeping. Type | None type checking is
updated to consistently use Optional[...]. Search and prompt_selection
methods are DRY'd out and the distinction of responsibilities is made
clearer for manipulating data and displaying it. Essentially, query is
handled in CLI menu, Gradebook actually searches its data and returns a
list, and a helper method displays that list of results for user
selection. Finally, the is_enrolled and is_archived boolean flags are
standardized to an is_active attribute, which will also be used by
Assignments (and perhaps Submissions). This should help standardize the
patterns that rely on activity status, as well as provide opportunities
for generalized function in the near future.