Refactor View Student and View Category
This commit restructures the view record portion of the Manage Students
and Manage Categories menu. Instead of immediately fracturing into the
different view options, it calls a sub-menu view_*_menu which has
options to view individual, active, inactive, or all records of the
relevant record type. This refactor also takes advantage of the recent
separation of concerns in the last commit, in which the Gradebook class
is responsible for filtering and supplying data, while the cli class is
responsbile for the output presentation. A helper class
sort_and_display_* helps DRY out the code a bit. A fuller abstraction
for these methods was deemed unnecessary at this point.