Add New Assignment
This adds a much more robust pattern for prompting record creation in
assignments_menu.py. Between the add_assignment and
prompt_new_assignment, there is a fuller structure for collecting data,
responding to bail-out attempts, previewing and confirming the record
data, trying any raise-possible conversions and object instantiation,
except to report any failures and gracefully exit, and ultimately
returning the object on successful initialization.

Missing still is the prompt_linked_category, which is called from
prompt_new_assignment and (hopefully) the final missing piece to this
menu's record creation.

If this all works well, it will be the blueprint for refactoring the
analagous methods in students_menu.py and categories_menu.py.