Formalize Add Records and formatters.py
The methods for creating new Students and Categories are brought in-line
with the standard established by the Assignments menu. Methods to search
for a record (of any type) and prompting for selection are extracted to
the menu_helpers.py to allow for sharing across menus, and formatting
methods are shelved into a formatters.py.