Edit Assignment and Assignment properties
Structures the editable fields for edit_assignment and the first few
methods (edit_assignment, edit_name...). Many assignment properties and
setters are added, and the class is updated to store the due_date as a
datetime object in an attribute, with multiple handlers to return string
formatted data representation.

A small assortment of formatters and menu_helpers is also added, as well
as some updated function names across classes for clarity and
specificity.