Remove Assignment and parallel methods
This implements the several methods involved in deleting, archiving, or
reactivating an Assignment record. It also updates the Category menu
somewhat and the Student menu more extensively to mirror the most recent
patterns involved in removing a record. A greater deal of flexibility,
more informational messaging and warnings, and closes a route where it
was possible to "archive" an archived record and thereby reactivate it
despite what the prompt was promising and reporting.