Add methods for adding new Submissions
The first stub of the Manage Submissions menu is largely implemented
here, adding a single submission. This required adding most of the
relevant methods for adding new submission and lays the ground work
nicely for the next add method - batch ad submissions by assignment.
Methods are included for linking assignment, linking student, preview
and confirm, handle existing submission, delete and create new
submissions, and so on. It does raise some new patterns that will need
to be cloned into the existing Manage Record menus, primarily separating
the search for a record from the editing a record, and factoring out the
functionality for prompting user input during data collection for record
creation.

Gradebook class is updated with a method for finding a submission by
assignment id and student id, as well as a method for checking whether
a submission already exists for a pair of assignment and student ids.

Tests are updated accordingly.