Update menu_helpers and View Submissions
The Submissions Menu is extended to include methods for viewing single
submissions or all submissions linked to a particular Assignment or
Student record. The methods prompting to link assignments and students
are also renamed for general use and moved to the bottom of the file.
Some repeated code finding those linked records is factored out as well.

Menu helpers is augmented with display_results and
display_sorted_results tailored to accept Submissions. The Submission
formatters required the gradebook to also be passed in as an argument in
order to retrieve the linked assignment and student objects. This extra
parameter precludes using the existing helpers, but otherwise they are
functionally the same.