Batch entry submissions by assignment
Adding the most elaborate menu option by far, batch entry of submissions
by assignment. The user chooses an assignment, a list of students who
have not yet submitted is populated and iterated over, and scores are
added for each student or skipped. Before committing submissions to the
gradebook data, the user has an opportunity to review the batch entered
submissions, make edits, or delete them, and then the skipped students
may also be reviewed and submissions added (this is TODO). The
relationship between functions was more complicated than anything else
so far, and although a few methods remain to be written, the work is
largely finished.