Add unsaved_changes flag and Submissions edit tree
Introduces a gradebook._unsaved_changes boolean attribute, as well as
methods for marking the gradebook dirty (unsaved) and clean (called
implicitly during the save function). This is not fully utilized yet,
but will make an easy flag for prompting saves at junction points only
if necessary. Currently, leaving almost any top-level menu prompts for
saving, even if there are no changes to be made.

The edit_submissions menu is fairly heavily refactored here. The
edit_queued_submissions revealed a huge overlap between edit_* functions
except for a single call to gradebook.save(). That prompted the idea to
extract that into the call site edit_submission, and the idea snowballed
from there. Add boolean return values from edit_* methods to indicate
whether there are unsaved changes, and allow the user to stay in the
batch edit flow without saving. Upon leaving the batch edit loop, if
there are unsaved changes, prompt to save now. If saving is deferred,
mark the gradebook dirty and continue along. The function scoped
unsaved_changes tracks the state of the current loop, but the gradebook
scoped unsaved_changes attribute allows the user to safely defer saving
without worrying about forgetting later on, exiting the program, and
losing changes they forgot to persists.

Furthermore, by extracting the save logic, I could reuse the edit_*
methods for queued submissions. They are simply called from a different
wrapper that ignore the boolean return values and does not prompt for
saving. There is no need to track data manipulation between creating a
submission and adding it to the Gradebook. The batch entry process is
going to prompt to save or discard all of the submissions no matter
what.

Overall, this was more refactoring than I anticipated, but it represents
a more airtight pattern that I can export to the earlier menus, and also
the idea of a Gradebook boolean dirty flag was appealing in earlier
design discussions already, so by adding it now I open the door to
future features rather nicely.