Add Review Skipped students to Batch Entry
This commit adds the review_skipped_students method for iterating over
any students skipped during batch entry. The user is given the
opportunity to either add a new submission for a selected student or
mark them exempt.

This also marks the first full implementation of the Submission Menu
according to the MVP design docs. I still have some formatting (== vs
is) issues to clean up, as well as make use of the newly added
gradebook.mark_duty() flag, but this represents a major arrival point in
the development of the CLI.