Refactor and Docstrings for Assignments Menu
This commit refactors and documents the assignments_menu.py file
according to the patterns and decisions previously applied to the
Submissions and Categories menus.

Search and select methods are extracted to the menu_helpers.py file, and
the menus now only have dispatch methods prompting the user to either
search by name  or select the record from a list, allowing menu specific
control such as "return None" where necessary.