Documentation and Refactoring the smaller files
Extends the refactoring and docstring creation applied to the four
larger menus to the smaller files in this commit. Documentation isn't
comprehensive - usually a file description and docstrings for more
involved methods, but properties and setters are ignored, as well as
to_dict and from_dict methods, for now.

In the cli/main.py, the create_gradebook() and load_gradebook()
reviews are deferred for now. They are important enough to warrant a
closer look once I implement Save As and other planned features.