Reorganize file structure, add Response object
Utils is moved into core/, cli menus are added to subdirectory
cli/menus/, and a new Response object is created in core/response.py.
This new object is meant to be the bundler object that reports results
from the Gradebook to the UI layer, with an eye towards standardization
and API life in the future. Will require major refactoring in the
existing menus, but the goal is to let the Gradebook be the interface
layer between the UI and the data models. Model objects have minimal
methods and make no assumptions about the Gradebook and the UI; the UI
asks the Gradebook to write data to a model, and the Gradebook knows how
to handle that request with the model's methods, then returns a Response
with success/failure, message, error, or data. Layer may access the data
models for read only operations exposed by properties.