Update Student attendance representation
This introduces a more nuanced way to represent student attendance.
Before, only absences were marked, and otherwise the student was
implicitly understood to be present. But in considering late, excused,
and unmarked statuses, it became clear that the current approach was
inadequate. This commit introduces an Enum class AttendanceReport that
allows for the finer detail. The Student attribute is changed from a set
of datetime.dates to a dictionary mapping datetime.dates to
AttendanceReport objects, and a suite of data accessors and manipulators
is added to the class.

In gradebook.py, the slow process of refactoring to include the new
Response model continues. In fact, the methods for generating attendance
reports by date and by student prompted the changes the Student class.
It's slow going, but the changes are methodical and also the
documentation is looking great. It will make returning to this project
several months later much easier when I'm ready to extend to GUI/API
layers.