Refactor Assignments menu for Response pattern
The Manage Assignments menu is fully refactored (as Students and
Categories before) to strictly dispatch through Gradebook and to handle
Response objects in return.

The Assignment class adds data validators for points_possible and
due_date inputs.

Gradebook is updated with no-op early exits for all write methods so
far, and also the full complement of write methods to go with the Manage
Assignments menu refactor.

Finally, the datetime import statement is untangled to allow for use of
datetime and date objects in the same file. Slightly more verbose, but
clearer across the entire program.