Implement Manage Course Schedule sub menu
First major steps after the CLI menu refactors of the last several days.
Returning to the Manage Attendance menu, which was interrupted to
overhaul Gradebook and introduce Response objects. This commit mostly
focuses on the Manage Course Schedule sub-menu, implementing methods to
add, batch-add, remove, and remove-all course dates to a schedule.
Several bespoke prompt and preview methods are also required, and the
generate recurring schedule method was especially fun. It does quite a
bit of heavy-lifting, but the end result is a smooth UX and easy way to
generate an entire course schedule with recurring dates and blackout
dates in one pass.

The other menus received very minor formatting and banner changes.