Starts work on Record Attendance flow
Makes significant steps towards the surprisingly complicated
record_attendance() method in the Manage Attendance menu.

Class date normalization is started with manual entry and enforcing course
schedule, but needs to still offer selection from a list of current
course schedule.

The main loop in record_attendance() is written in pseudocode.

The heavy-lifting today was the additional architecture required to
allow the user to stage changes before committing, and also the ability
to shift between the broad-stroke record_attendance() and the scalpel
accurate edit_attendance(). That necessitated a flexible menu of options
based on the current state of attendance. The end result was a new
in-module class GatewayState to present a snapshot generated newly with
each loop, a GatewayChoice to indicate the menu choice and dispatch
accordingly, and a new helper class AttendanceStager in the core modules
to keep track of changes.

This was much more complicated than I anticipated and unlike anything
else in the program, but I think we these classes and models in place,
I'm well suited to implement the UI flow and Gradebook write operations
tomorrow.