Refactor formatter modules
To solve an import loop caused by the Gradebook formatting dates for the
benefit of Response details, the cli/formatters.py module was divided
along the line of strict formatters and
model-dependent/gradebook-reading formatters. The former live within
core/formatters.py and imported as formatters, the latter live within
cli/model_formatters.py and imported as model_formatters. All modules
that previously included import cli.formatters as formatters are
refactored to accomodate this change.

In rewriting the formatters, it was also an opportunity to gently
reconsider model preview layouts, as well as update the formatters to
properly handle Gradebook Response objects after read-only operations.