Implement Edit Attendance flow
This adds the necessary methods to implement the edit attendance suite.
The user is prompted to either edit attendance records by date or by
student, and forks accordingly. If the edit_by_date or edit_by_student
methods are called without a date/student as argument, user is prompted
to resolve by search or selection from list. However, a student or date
can be provided to fork into the edit sequence from other callsites
where that data is already resolved (such as record_attendance, date has
already been selected at that point). Then the user enters a loop to
select the complementary data (date or student) and prompted to assign a
new value for the attendance record. The user returns to the loop to
allow continual editing for that student/date without needing to climb
all the way back.