Complete the Manage Attendance mene suite
This commit marks the full implementation of the Manage Attendance CLI
menu. All methods for adding, editing, removing, and recording
attendance are written and documented, as well as helper, formatting,
and Gradebook level methods required to support the menu. This also
marks the last element of record-keeping functionality! All that
remains now is the reporting suite (and tests ... so many tests).