Test coverage for gradebook.py
Tests written for all methods with expected use cases. The next step is
to augment these tests with edge cases, but at least the intended
functionality is covered and working as intended. One big bug was caught
as a result (returning copies of the Gradebook attribute dictionaries to
internal write methods, which meant nothing was persisting!).